--- a/package.json
+++ b/package.json
@@ -21,6 +21,7 @@
     "angular2": "2.0.0-beta.0",
     "es6-promise": "^3.0.2",
     "es6-shim": "^0.33.3",
+    "fs-extra": "^0.26.5",
     "reflect-metadata": "0.1.2",
     "rxjs": "5.0.0-beta.0",
     "systemjs": "0.19.4",