Broccoli itself isn't properly checking diffs and as such takes a long
time to rebuild the app. Angular built a diffing broccoli wrapper that
does that effectively.

With this PR now the slowest tree is the TS2Dart one.

## Before this change:

Hot start:

| Slowest Trees                                 | Total
|----------------------------------------------|---------------------|
|DiffingTSCompiler                             | 1605ms|
|TSToDartTranspiler                            | 855ms|
|  SassCompiler                                  | 311ms
|    SassCompiler                                  | 293ms
----
Refresh:

| Slowest Trees                                 | Total
|----------------------------------------------|---------------------
|DiffingTSCompiler                             | 1169ms
|TSToDartTranspiler                            | 832ms

----
## With the new instances

Hot start:

|Slowest Trees                                 | Total
|----------------------------------------------|---------------------
|TSToDartTranspiler                            | 692ms
|DiffingTSCompiler                             | 684ms
|AutoprefixerFilter                            | 90ms

----
Refresh:

| Slowest Trees                                 | Total
|----------------------------------------------|---------------------
|TSToDartTranspiler                            | 526ms
|DiffingTSCompiler                             | 61ms