@hansl 2016-02-02
Apparently this doesn't work better at all, since the `angular-cli` import does the same. I was just skipping over a file copying step. Closing this.